+++ src/addons/send2ue/core/io/__init__.py
@@ -1,6 +1,5 @@
-from . import fbx_b3, fbx_b4
+from . import fbx_b4
 
 __all__ = [
-    'fbx_b3',
     'fbx_b4'
 ]